justerat lite grann

+++ SDG Admin/src/SQLHanterare/AnstalldHanterare.java
@@ -166,7 +166,7 @@ public class AnstalldHanterare {
   
        }
        
-      catch (Exception e) {
+      catch (InfException e) {
           e.printStackTrace();
           return new ArrayList<>();
        
@@ -185,10 +185,20 @@ public class AnstalldHanterare {
    
    
    public boolean laggTillAnstalld(String fornamn, String efternamn, String adress, String epost, String telefon, String anstallningsdatum)
-    {
+    
+   {
         
-        try
+       
         {
+        
+          if (fornamn == null || efternamn == null || adress == null || epost == null || telefon == null || anstallningsdatum == null || fornamn.isEmpty() || efternamn.isEmpty() || adress.isEmpty() || epost.isEmpty() ||  telefon.isEmpty() || anstallningsdatum.isEmpty()) {
+          
+           System.out.println("Du har glömt att fylla i ett eller fler fält. Anställd kan inte läggas till");
+           return false;
+           
+       }
+        
+        try {
              
          String losenord = UUID.randomUUID().toString().substring(0, 9);
          String aid = UUID.randomUUID().toString();
@@ -202,7 +212,7 @@ public class AnstalldHanterare {
             return true;
         }
         
-        catch (Exception e)
+        catch (InfException e)
         {
             e.printStackTrace();
             return false;
@@ -211,18 +221,27 @@ public class AnstalldHanterare {
         
    
     }
-
+   }
 
 // koderna under avsedda för att ändra mina uppgifter.
 
 public boolean andraEpost(String aid, String nyEpost)
 {
+    {
+        if (aid == null || nyEpost == null || aid.isEmpty() || nyEpost.isEmpty()) {
+            System.out.println("aid eller epost får inte vara tom");
+            return false;
+        }
+    }
+           
+    
+    
     try {
         String ePost = "UPDATE anstalld SET epost = '" + nyEpost + "' WHERE aid = '" + aid + "'";
         idb.update(ePost);
         return true;
 } 
-    catch (Exception e) {
+    catch (InfException e) {
     e.printStackTrace();
     return false; 
 }
@@ -231,12 +250,22 @@ public boolean andraEpost(String aid, String nyEpost)
 
     public boolean andraLosenord(String aid, String nyttLosenord)
 {
+    
+     {
+        if (aid == null || nyttLosenord == null || aid.isEmpty() || nyttLosenord.isEmpty()) {
+            System.out.println("aid eller losenird får inte vara tom");
+            return false;
+        }
+    }
+    
+    
+    
     try {
           String losenord = "UPDATE anstalld SET losenord = '" + nyttLosenord + "' WHERE aid = '" + aid + "'";
           idb.update(losenord);
           return true;
 }
-    catch (Exception e) {
+    catch (InfException e) {
     
         e.printStackTrace();
         return false;
@@ -250,6 +279,14 @@ public boolean andraEpost(String aid, String nyEpost)
 
 public boolean andraFornamn(String aid, String nyttFornamn)
 {
+   {
+        if (aid == null || nyttFornamn == null || aid.isEmpty() || nyttFornamn.isEmpty()) {
+            System.out.println("aid eller förnamn får inte vara tom");
+            return false;
+        }
+    }
+    
+    
     try {
 
            String fornamn = "UPDATE anstalld SET fornamn = '" + nyttFornamn + "' WHERE aid = '" + aid + "'";
@@ -257,7 +294,7 @@ public boolean andraFornamn(String aid, String nyttFornamn)
            return true;
 }
     
-        catch (Exception e)
+        catch (InfException e)
 
 {
         e.printStackTrace();
@@ -271,6 +308,15 @@ public boolean andraFornamn(String aid, String nyttFornamn)
 
 public boolean andraEfternamn(String aid, String nyttEfternamn)
 {
+     {
+        if (aid == null || nyttEfternamn == null || aid.isEmpty() || nyttEfternamn.isEmpty()) {
+            System.out.println("aid eller efternamn får inte vara tom");
+            return false;
+        }
+    }
+    
+    
+    
     try {
 
            String efternamn = "UPDATE anstalld SET efternamn = ' " + nyttEfternamn + " ' WHERE aid = ' " + aid + "'";
@@ -278,7 +324,7 @@ public boolean andraEfternamn(String aid, String nyttEfternamn)
            return true;
 }
     
-        catch (Exception e)
+        catch (InfException e)
 
 {
         e.printStackTrace();
@@ -290,6 +336,13 @@ public boolean andraEfternamn(String aid, String nyttEfternamn)
 
 public boolean andraAdress(String aid, String nyAdress)
 {
+    {
+        if (aid == null || nyAdress == null || aid.isEmpty() || nyAdress.isEmpty()) {
+            System.out.println("aid eller adress får inte vara tom");
+            return false;
+        }
+    }
+    
     try {
 
            String adress = "UPDATE anstalld SET adress = ' " + nyAdress + " ' WHERE aid = ' " + aid + "'";
@@ -297,7 +350,7 @@ public boolean andraAdress(String aid, String nyAdress)
            return true;
 }
     
-        catch (Exception e)
+        catch (InfException e)
 
 {
         e.printStackTrace();
@@ -308,17 +361,31 @@ public boolean andraAdress(String aid, String nyAdress)
 
 }
 
-// radera anställd
+/**
+ * Denna klass raderar en anställd från databasen baserat på deras aid.
+ * Valdideringen sker genom att i if-satsen kollar den att aid inte
+ * är null eller tom
+ * 
+ */
 
     public boolean taBortAnstalld (Anstalld a)
 {
 try {
+    
+    String aid = a.getAid();
+    
+    if (aid == null || aid.isEmpty()) {
+        System.out.println("Aid är tom");
+        return false;
+    }
+    
     String taBort = "DELETE FROM anstalld WHERE aid = '" + a.getAid() + "'";
     idb.delete(taBort);
-    return true;
+    System.out.println("Anställd borttagen: "  + a.getFornamn() + " " + a.getEfternamn());
+    return true;    
 }
 
-catch (Exception e) {
+catch (InfException e) {
     
     e.printStackTrace();
